requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
appnope==0.1.3
    # via ipython
asttokens==2.1.0
    # via stack-data
backcall==0.2.0
    # via ipython
cloudpickle==2.2.0
    # via gym
decorator==5.1.1
    # via ipython
executing==1.2.0
    # via stack-data
gym==0.26.2
    # via mdptoolbox-hiive
gym-notices==0.0.8
    # via gym
importlib-metadata==5.0.0
    # via gym
ipython==8.6.0
    # via mdptoolbox-hiive
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
mdptoolbox-hiive @ git+https://github.com/BigTMiami/hiivemdptoolbox.git
    # via -r requirements.in
networkx==2.8.8
    # via mdptoolbox-hiive
numpy==1.23.4
    # via
    #   gym
    #   mdptoolbox-hiive
    #   pandas
    #   scipy
pandas==1.5.1
    # via -r requirements.in
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.32
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pydot==1.4.2
    # via mdptoolbox-hiive
pygments==2.13.0
    # via ipython
pyparsing==3.0.9
    # via pydot
python-dateutil==2.8.2
    # via pandas
pytz==2022.6
    # via pandas
scipy==1.9.3
    # via mdptoolbox-hiive
six==1.16.0
    # via python-dateutil
stack-data==0.6.0
    # via ipython
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.5
    # via prompt-toolkit
zipp==3.10.0
    # via importlib-metadata
